refactor(experience): merge the Dashboard into Health — one observability home (#314)
Health and Dashboard both showed 'the state of your system' and overlapped on
debt + connectivity. Centralize into one Health mode: it now also renders the
connectivity + 'today' panels (reusing buildKnowledgeDashboard, skipping the
debt panel since Health's drill-down is richer), then debt + balance + the
orphan/dead-end lists. Delete the Dashboard mode + KnowledgeDashboardRenderer;
show-knowledge-dashboard + the zettelflow-knowledge-dashboard view redirect to
Health via legacyTargets. Prune orphaned dashboard chrome/toolkit i18n + the
docs link. Tests + docs + README (two overlapping bullets merged into one).
