+++ src/architecture/components/core/surface/HealthSurfaceView.ts
@@ -1,14 +1,13 @@
 import { ModeHostView } from "./ModeHostView";
 import { KnowledgeModeRenderer } from "./KnowledgeModeRenderer";
 import { SlipboxHealthRenderer } from "architecture/components/core/slipboxHealth/SlipboxHealthRenderer";
-import { KnowledgeDashboardRenderer } from "architecture/components/core/knowledgeDashboard/KnowledgeDashboardRenderer";
 import { EvolutionTimelineRenderer } from "architecture/components/core/timeline/EvolutionTimelineRenderer";
 import { ThinkingHeatmapRenderer } from "architecture/components/core/thinkingHeatmap/ThinkingHeatmapRenderer";
 
 /**
  * The **Health** surface (#272) — one destination for the state of your slip-box, with modes:
- * Health (debt + balance) · Dashboard (ops console) · Timeline (a note's evolution) · Momentum (the
- * development heatmap). Each mode reuses the retired view's renderer verbatim.
+ * Health (connectivity · today · debt · balance · orphans/dead-ends — the ops console merged in,
+ * #314) · Timeline (a note's evolution) · Momentum (the development heatmap).
  */
 export class HealthSurfaceView extends ModeHostView {
     getViewType(): string {
@@ -21,8 +20,6 @@ export class HealthSurfaceView extends ModeHostView {
 
     protected createRenderer(modeId: string, container: HTMLElement): KnowledgeModeRenderer {
         switch (modeId) {
-            case "dashboard":
-                return new KnowledgeDashboardRenderer(container, this.app);
             case "timeline":
                 return new EvolutionTimelineRenderer(container, this.app);
             case "momentum":
