+++ src/architecture/components/core/surface/surfaceRegistry.ts
@@ -37,7 +37,6 @@ export const SURFACES: readonly Surface[] = [
         titleKey: "surface_health_title",
         modes: [
             { id: "health", sourceView: "zettelflow-slipbox-health", labelKey: "surface_mode_health" },
-            { id: "dashboard", sourceView: "zettelflow-knowledge-dashboard", labelKey: "surface_mode_dashboard" },
             { id: "timeline", sourceView: "zettelflow-evolution-timeline", labelKey: "surface_mode_timeline" },
             { id: "momentum", sourceView: "zettelflow-thinking-heatmap", labelKey: "surface_mode_momentum" },
         ],
