+++ test/architecture/components/core/surface/surfaceRegistry.test.ts
@@ -14,13 +14,13 @@ describe("surface registry (#272, AC-1/AC-6)", () => {
     });
 
     it("hosts the retired views as modes, each exactly once (net-new modes have no sourceView)", () => {
-        // 10 retired views survive as modes; the 2 retired Graph views (knowledge-map/concept-nav) now
-        // redirect to the 3D mode via legacyTargets rather than being modes (#280 — Graph is 3D-only).
+        // 9 retired views survive as modes; 3 now redirect via legacyTargets rather than being modes:
+        // the 2 Graph views (knowledge-map/concept-nav → 3D, #280) and knowledge-dashboard (→ Health, #314).
         const sources = SURFACES.flatMap((s) => s.modes.map((m) => m.sourceView)).filter(
             (s): s is string => s !== undefined
         );
-        expect(sources).toHaveLength(10);
-        expect(new Set(sources).size).toBe(10);
+        expect(sources).toHaveLength(9);
+        expect(new Set(sources).size).toBe(9);
     });
 
     it("locateSourceView resolves every retired source to its (surface, mode); defaultMode is the first mode", () => {
