+++ test/config/surfaceLocale.test.ts
@@ -10,8 +10,8 @@ describe("surface i18n keys (#272)", () => {
     const enMap = en as Record<string, string>;
     const esMap = es as Record<string, string>;
 
-    it("references 4 titles + 12 mode labels (Home adds Cultivate, #309)", () => {
-        expect(KEYS).toHaveLength(16);
+    it("references 4 titles + 11 mode labels (Dashboard merged into Health, #314)", () => {
+        expect(KEYS).toHaveLength(15);
     });
 
     it("defines every surface/mode key in both en and es, non-empty", () => {
